Also rename gui/dialogs/game_version.hpp
Necessary for MSVC builds.
