+++ projectfiles/VC14/wesnoth.vcxproj
@@ -1681,7 +1681,7 @@
       <ObjectFileName Condition="'$(Configuration)|$(Platform)'=='Test_Debug|Win32'">$(IntDir)Gui\Dialogs\</ObjectFileName>
       <ObjectFileName Condition="'$(Configuration)|$(Platform)'=='Test_Release|Win32'">$(IntDir)Gui\Dialogs\</ObjectFileName>
     </ClCompile>
-    <ClCompile Include="..\..\src\gui\dialogs\game_version.cpp">
+    <ClCompile Include="..\..\src\gui\dialogs\game_version_dialog.cpp">
       <ObjectFileName Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">$(IntDir)Gui\Dialogs\</ObjectFileName>
       <ObjectFileName Condition="'$(Configuration)|$(Platform)'=='ReleaseDEBUG|Win32'">$(IntDir)Gui\Dialogs\</ObjectFileName>
       <ObjectFileName Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">$(IntDir)Gui\Dialogs\</ObjectFileName>
@@ -3728,7 +3728,7 @@
     <ClInclude Include="..\..\src\gui\dialogs\game_save.hpp" />
     <ClInclude Include="..\..\src\gui\dialogs\game_stats.hpp" />
     <ClInclude Include="..\..\src\gui\dialogs\game_ui.hpp" />
-    <ClInclude Include="..\..\src\gui\dialogs\game_version.hpp" />
+    <ClInclude Include="..\..\src\gui\dialogs\game_version_dialog.hpp" />
     <ClInclude Include="..\..\src\gui\dialogs\help_browser.hpp" />
     <ClInclude Include="..\..\src\gui\dialogs\hotkey_bind.hpp" />
     <ClInclude Include="..\..\src\gui\dialogs\ingame_ui_base.hpp" />
