+++ source_lists/wesnoth
@@ -195,7 +195,7 @@ gui/dialogs/game_load.cpp
 gui/dialogs/game_save.cpp
 gui/dialogs/game_stats.cpp
 gui/dialogs/game_ui.cpp
-gui/dialogs/game_version.cpp
+gui/dialogs/game_version_dialog.cpp
 gui/dialogs/gamestate_inspector.cpp
 gui/dialogs/help_browser.cpp
 gui/dialogs/hotkey_bind.cpp
