+++ src/gui/dialogs/game_version_dialog.cpp
@@ -14,7 +14,7 @@
 
 #define GETTEXT_DOMAIN "wesnoth-lib"
 
-#include "gui/dialogs/game_version.hpp"
+#include "gui/dialogs/game_version_dialog.hpp"
 
 #include "build_info.hpp"
 #include "desktop/clipboard.hpp"
